BREAKING CHANGES: modificando a estrutura do codigo para que as rotas possam ser modularizadas
