+++ app/extensions.py
@@ -1,17 +1,15 @@
-from app import app
-from flask_marshmallow import Marshmallow
-from flask_migrate import Migrate
-from flask_sqlalchemy import SQLAlchemy
-from app.config.app_config import *
-from sqlalchemy.types import TypeDecorator, CHAR
-from sqlalchemy.dialects.postgresql import UUID
 import uuid
 from datetime import datetime
+from flask_migrate import Migrate
+from flask_sqlalchemy import SQLAlchemy
+from flask_marshmallow import Marshmallow
+from sqlalchemy.dialects.postgresql import UUID
+from sqlalchemy.types import TypeDecorator, CHAR
 
 
-db = SQLAlchemy(app)
-migrate = Migrate(app, db)
-ma = Marshmallow(app)
+db = SQLAlchemy()
+migrate = Migrate()
+ma = Marshmallow()
 
 class GUID(TypeDecorator):
     """Platform-independent GUID type.
