+++ app/models/tables/category.py
@@ -1,17 +1,12 @@
-from app.config.app_config import *
-from app.config.db_config import *
-from app.models.tables.movie import Movie
-from datetime import datetime
+from app.extensions import Base
 from sqlalchemy.orm import Mapped, mapped_column, relationship, backref
-from sqlalchemy import String, DateTime
+from sqlalchemy import String
 from uuid import uuid4
 
 
-class Category(db.Model):
+class Category(Base):
     id: Mapped[str] = mapped_column(String, primary_key=True, default=lambda: str(uuid4()))
     name: Mapped[str] = mapped_column(String(45), unique=True, nullable=False)
-    created_at: Mapped[DateTime] = mapped_column(DateTime, nullable=False, default=datetime.utcnow)
-    updated_at: Mapped[DateTime] = mapped_column(DateTime, nullable=False, default=datetime.utcnow, onupdate=datetime.utcnow)
     movies = relationship('Movie', backref='category') 
     
     def __init__(self, name:str):
