+++ run.py
@@ -1,6 +1,6 @@
-from app import app
+from app import create_app
 
-
+app = create_app()
 
 if __name__ == '__main__':
     app.run(port=5000, debug=True)
